seperate cli and cmd main package
