+++ cmd/bbolt/main_test.go
@@ -18,7 +18,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	bolt "go.etcd.io/bbolt"
-	main "go.etcd.io/bbolt/cmd/bbolt"
+	"go.etcd.io/bbolt/cmd/cli"
 	"go.etcd.io/bbolt/internal/guts_cli"
 )
 
@@ -441,7 +441,7 @@ func (b *ConcurrentBuffer) String() string {
 
 // Main represents a test wrapper for main.Main that records output.
 type Main struct {
-	*main.Main
+	*cli.Main
 	Stdin  ConcurrentBuffer
 	Stdout ConcurrentBuffer
 	Stderr ConcurrentBuffer
@@ -449,7 +449,7 @@ type Main struct {
 
 // NewMain returns a new instance of Main.
 func NewMain() *Main {
-	m := &Main{Main: main.NewMain()}
+	m := &Main{Main: cli.NewMain()}
 	m.Main.Stdin = &m.Stdin
 	m.Main.Stdout = &m.Stdout
 	m.Main.Stderr = &m.Stderr
