+++ cmd/cli/command_surgery.go
@@ -1,4 +1,4 @@
-package main
+package cli
 
 import (
 	"errors"
