+++ cmd/cli/command_surgery_freelist.go
@@ -1,4 +1,4 @@
-package main
+package cli
 
 import (
 	"errors"
