+++ cmd/cli/command_surgery_freelist_test.go
@@ -1,4 +1,4 @@
-package main_test
+package cli_test
 
 import (
 	"path/filepath"
@@ -8,7 +8,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	bolt "go.etcd.io/bbolt"
-	main "go.etcd.io/bbolt/cmd/bbolt"
+	main "go.etcd.io/bbolt/cmd/cli"
 	"go.etcd.io/bbolt/internal/btesting"
 	"go.etcd.io/bbolt/internal/common"
 )
