+++ cmd/cli/page_command.go
@@ -1,4 +1,4 @@
-package main
+package cli
 
 import (
 	"bytes"
